# smart_home/smart home/backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能家居综合管理系统后端",
    "description": "智能家居综合管理系统 Python FastAPI 后端接口服务。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/homes/{home_id}/members/invite": {
      "post": {
        "tags": [
          "成员管理"
        ],
        "summary": "Invite Member",
        "operationId": "invite_member_api_homes__home_id__members_invite_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberInviteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/homes/{home_id}/members/apply": {
      "post": {
        "tags": [
          "成员管理"
        ],
        "summary": "Apply Member",
        "operationId": "apply_member_api_homes__home_id__members_apply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberApplyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/homes/{home_id}/members/{member_id}/accept": {
      "post": {
        "tags": [
          "成员管理"
        ],
        "summary": "Accept Invitation",
        "operationId": "accept_invitation_api_homes__home_id__members__member_id__accept_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/homes/{home_id}/members/{member_id}/approve": {
      "post": {
        "tags": [
          "成员管理"
        ],
        "summary": "Approve Member",
        "operationId": "approve_member_api_homes__home_id__members__member_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/homes/{home_id}/members": {
      "get": {
        "tags": [
          "成员管理"
        ],
        "summary": "List Members",
        "operationId": "list_members_api_homes__home_id__members_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/homes/{home_id}/members/{member_id}/permission": {
      "put": {
        "tags": [
          "成员管理"
        ],
        "summary": "Update Permission",
        "operationId": "update_permission_api_homes__home_id__members__member_id__permission_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberPermissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/homes/{home_id}/members/{member_id}": {
      "delete": {
        "tags": [
          "成员管理"
        ],
        "summary": "Remove Member",
        "operationId": "remove_member_api_homes__home_id__members__member_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/homes/{home_id}/rooms": {
      "post": {
        "tags": [
          "房间管理"
        ],
        "summary": "新增房间",
        "operationId": "create_room_api_homes__home_id__rooms_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Room Api Homes  Home Id  Rooms Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "房间管理"
        ],
        "summary": "查询家庭空间下的房间列表",
        "operationId": "list_rooms_api_homes__home_id__rooms_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Rooms Api Homes  Home Id  Rooms Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_id}": {
      "put": {
        "tags": [
          "房间管理"
        ],
        "summary": "修改房间",
        "operationId": "update_room_api_rooms__room_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Room Api Rooms  Room Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "房间管理"
        ],
        "summary": "删除房间",
        "operationId": "delete_room_api_rooms__room_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Room Api Rooms  Room Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms": {
      "get": {
        "tags": [
          "房间管理"
        ],
        "summary": "兼容旧接口：查询房间列表",
        "operationId": "list_rooms_compat_api_rooms_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "家庭空间编号",
              "title": "Home Id"
            },
            "description": "家庭空间编号"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Rooms Compat Api Rooms Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "房间管理"
        ],
        "summary": "兼容旧接口：新增房间",
        "operationId": "create_room_compat_api_rooms_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreateCompat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Room Compat Api Rooms Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberApplyRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "MemberApplyRequest"
      },
      "MemberApproveRequest": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved",
            "default": true
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "MEMBER"
          }
        },
        "type": "object",
        "title": "MemberApproveRequest"
      },
      "MemberInviteRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "MEMBER"
          },
          "expire_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expire At"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "MemberInviteRequest"
      },
      "MemberPermissionRequest": {
        "properties": {
          "role": {
            "type": "string",
            "pattern": "^(OWNER|MEMBER|GUEST|MAINTAINER)$",
            "title": "Role"
          },
          "expire_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expire At"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "MemberPermissionRequest"
      },
...
      "RoomCreate": {
        "properties": {
          "room_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Room Name",
            "description": "房间名称"
          },
          "room_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Type",
            "description": "房间类型"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "description": "备注"
          }
        },
        "type": "object",
        "required": [
          "room_name"
        ],
        "title": "RoomCreate"
      },
      "RoomCreateCompat": {
        "properties": {
          "room_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Room Name",
            "description": "房间名称"
          },
          "room_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Type",
            "description": "房间类型"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "description": "备注"
          },
          "home_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Home Id",
            "description": "家庭空间编号"
          }
        },
        "type": "object",
        "required": [
          "room_name",
          "home_id"
        ],
        "title": "RoomCreateCompat"
      },
      "RoomUpdate": {
        "properties": {
          "room_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Room Name",
            "description": "房间名称"
          },
          "room_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Type",
            "description": "房间类型"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "description": "备注"
          }
        },
        "type": "object",
        "required": [
          "room_name"
        ],
        "title": "RoomUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}